--- src/main/java/org/example/demo6/Controller/HomePageController.java
@@ -1,4 +1,0 @@
-package org.example.demo6.Controller;
-
-public class HomePageController {
-}
