--- src/main/java/org/example/demo6/Controller/PageDiscovery.java
@@ -1,5 +1,0 @@
-package org.example.demo6.Controller;
-
-public class PageDiscovery {
-    
-}
