+++ src/main/java/org/example/demo6/HomePageController.java
@@ -1,0 +1,4 @@
+package org.example.demo6;
+
+public class HomePageController {
+}
