+++ src/main/java/org/example/demo6/LoginController.java
@@ -1,4 +1,4 @@
-package org.example.demo6.Controller;
+package org.example.demo6;
 
 import com.jfoenix.controls.JFXButton;
 import javafx.event.ActionEvent;
@@ -6,7 +6,7 @@ import javafx.fxml.FXML;
 import javafx.scene.control.Button;
 import javafx.scene.control.PasswordField;
 import javafx.scene.control.TextField;
-import org.example.demo6.DBUltis;
+
 
 
 public class LoginController {
