+++ src/main/java/org/example/demo6/MainSceneClass.java
@@ -1,4 +1,4 @@
-package org.example.demo6.Controller;
+package org.example.demo6;
 
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
@@ -10,8 +10,6 @@ import javafx.scene.control.*;
 import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.VBox;
 import javafx.stage.Stage;
-import org.example.demo6.DBUltis;
-import org.example.demo6.apiGoogleBooks;
 
 import java.io.IOException;
 import java.util.Optional;
