+++ src/main/java/org/example/demo6/PageDiscovery.java
@@ -1,0 +1,5 @@
+package org.example.demo6;
+
+public class PageDiscovery {
+    
+}
