+++ src/main/java/org/example/demo6/apiGoogleBooks.java
@@ -40,7 +40,7 @@ public class apiGoogleBooks {
                 Volume.VolumeInfo volumeInfo = volume.getVolumeInfo();
                 System.out.println(volumeInfo.getTitle() + " by " + volumeInfo.getAuthors()
                         + " - " + volumeInfo.getPublishedDate() + " - " + volumeInfo.getPageCount() + " pages"
-                        + " - " + volumeInfo.getCategories() + " - " + volumeInfo.getImageLinks());
+                        + " - " + volumeInfo.getCategories() );
 
             }
         } else {
