+++ src/main/resources/View/ForgotPassword.fxml
@@ -5,7 +5,7 @@
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.text.Text?>
 
-<AnchorPane prefHeight="534.0" prefWidth="792.0" xmlns="http://javafx.com/javafx/23.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.example.demo6.Controller.ForgotPasswordController">
+<AnchorPane prefHeight="534.0" prefWidth="792.0" xmlns="http://javafx.com/javafx/23.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.example.demo6.ForgotPasswordController">
    <children>
       <AnchorPane fx:id="ap" layoutX="159.0" layoutY="304.0" prefHeight="229.0" prefWidth="612.0" visible="false">
          <children>
