+++ src/main/resources/View/PageDiscovery.fxml
@@ -8,7 +8,7 @@
 
 <AnchorPane xmlns="http://javafx.com/javafx"
             xmlns:fx="http://javafx.com/fxml"
-            fx:controller="org.example.demo6.Controller.LoginController"
+            fx:controller="org.example.demo6.LoginController"
             prefHeight="400.0" prefWidth="600.0">
 
 </AnchorPane>
