+++ src/main/resources/View/PageUpload.fxml
@@ -1,24 +1,11 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
 <?import javafx.scene.control.Button?>
-<?import javafx.scene.control.DatePicker?>
-<?import javafx.scene.control.TextField?>
-<?import javafx.scene.image.ImageView?>
 <?import javafx.scene.layout.AnchorPane?>
 
-<AnchorPane prefHeight="558.0" prefWidth="703.0" xmlns="http://javafx.com/javafx/23.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.example.demo6.Controller.PageUploadController">
+
+<AnchorPane prefHeight="490.0" prefWidth="703.0" xmlns="http://javafx.com/javafx/23" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.example.demo6.PageUploadController">
    <children>
-      <Button fx:id="uploadButton" layoutX="39.0" layoutY="426.0" mnemonicParsing="false" text="Add text book" />
-      <TextField fx:id="Tittle" layoutX="305.0" layoutY="67.0" prefHeight="25.0" prefWidth="209.0" promptText="Tittle" />
-      <TextField fx:id="Authors" layoutX="305.0" layoutY="101.0" prefHeight="25.0" prefWidth="209.0" promptText="Authors" />
-      <TextField fx:id="Publisher" layoutX="305.0" layoutY="152.0" prefHeight="25.0" prefWidth="209.0" promptText="Publisher" />
-      <TextField fx:id="Categories" layoutX="305.0" layoutY="237.0" prefHeight="25.0" prefWidth="209.0" promptText="Categories" />
-      <TextField fx:id="Language" layoutX="305.0" layoutY="279.0" prefHeight="25.0" prefWidth="209.0" promptText="Language (Vietnamese, English, ..)" />
-      <TextField fx:id="Description" layoutX="305.0" layoutY="330.0" prefHeight="56.0" prefWidth="209.0" promptText="Description" />
-      <Button fx:id="AudioBook" layoutX="472.0" layoutY="426.0" mnemonicParsing="false" text="Add audio book (if have)" />
-      <Button fx:id="SelectCover" layoutX="292.0" layoutY="426.0" mnemonicParsing="false" text="Book cover" />
-      <DatePicker fx:id="PublishedDate" layoutX="305.0" layoutY="191.0" prefHeight="25.0" prefWidth="209.0" promptText="Published date" />
-      <ImageView fx:id="BookCover" fitHeight="257.0" fitWidth="203.0" layoutX="39.0" layoutY="88.0" pickOnBounds="true" preserveRatio="true" />
-      <Button fx:id="Confirm" layoutX="304.0" layoutY="493.0" mnemonicParsing="false" text="Confirm" />
+      <Button fx:id="uploadButton" layoutX="261.0" layoutY="116.0" mnemonicParsing="false" text="Button" />
    </children>
 </AnchorPane>
