+++ src/main/resources/View/TopBooks.fxml
@@ -4,7 +4,7 @@
 <?import javafx.scene.control.Label?>
 <?import javafx.scene.layout.AnchorPane?>
 
-<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/22" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.example.demo6.Controller.HelloController">
+<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/22" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.example.demo6.HelloController">
    <children>
       <Label layoutX="213.0" layoutY="168.0" text="This is top book" />
       <Button layoutX="449.0" layoutY="66.0" mnemonicParsing="false" onAction="#returnHome" text="&lt;-" />
